+++ src/confidential_data_pipeline/prepare_incidence_data.R
@@ -14,52 +14,25 @@ library(metR)
 library(lubridate)
 library(ggpubr)
 
-# set to directory
-indir <- "~/git/phyloflows"
+# do you want to restrict the analysis to the 30 communities continuously surveyed?
+restrict_to_30_comms <- F
 
-# set up path to data
-indir_deepsequencedata <- file.path("~/Box\ Sync",
-                                    "2019",
-                                    "ratmann_pangea_deepsequencedata",
-                                    "live")
-indir_deepsequence_analyses <- file.path("~/Box\ Sync",
-                                         "2021",
-                                         "ratmann_deepseq_analyses",
-                                         "live")
+# paths
+gitdir <- here()
+source(file.path(gitdir, "config.R"))
 
-# set path to save results
-outdir <- file.path(indir_deepsequence_analyses,
-                    "PANGEA2_RCCS",
-                    "incidence_rate_inland")
+c(  file.path.flow.614,
+    file.path.hiv.614,
+    file.path.flow,
+    file.path.hiv.1518,
+    file.path.flow_19,
+    file.path.hiv_19) |> file.exists() |> all() |> stopifnot()
 
-# round 6 to 18: only use round 6 to 14
-path_flow_614 <- file.path(indir_deepsequencedata,
-                           "RCCS_data_estimate_incidence_inland_R6_R18",
-                           "220903",
-                           "verif_1.dta")
-path_hiv_614 <- file.path(indir_deepsequencedata,
-                               "RCCS_data_estimate_incidence_inland_R6_R18",
-                               "220903",
-                               "hivincidence_1.dta")
-
-# round 15 to 18
-path_flow_1518 <- file.path(indir_deepsequencedata,
-                            "RCCS_R15_R18",
-                            "FlowR15_R18_VoIs_220129.csv")
-path_hiv_1518 <- file.path(indir_deepsequencedata,
-                           "RCCS_R15_R18",
-                           "HIV_R15_R18_VOIs_220129.csv")
-
-# round 19
-path_flow_19 <- file.path(indir_deepsequencedata,
-                               "R019_VoIs",
-                               "Flow_R019_VoIs_220607.csv")
-path_hiv_19 <- file.path(indir_deepsequencedata,
-                         "R019_VoIs",
-                         "HIV_R019_VOIs_220607.csv")
+# save intermediary results not on Zolodo
+outdir <- file.path(indir.deepsequence_analyses, "PANGEA2_RCCS", "incidence_rate_inland")
 
 # function
-source(file.path(indir, "functions", "incidence_rate_estimation_functions.R"))
+source(file.path('R', "functions_incidence_rate", "incidence_rate_estimation_functions.R"))
 
 # utils
 rounds_group_1 <- c("R006","R007", "R008", "R009", "R010",
@@ -75,7 +48,6 @@ rounds_numeric_group_2 <- df_round[visit %in% rounds_group_2, round_numeric]
 rounds_numeric_group_3 <- df_round[visit == rounds_group_3, round_numeric]
 
 
-
 ##############
 
 # LOAD DATA #
@@ -83,14 +55,14 @@ rounds_numeric_group_3 <- df_round[visit == rounds_group_3, round_numeric]
 ##############
 
 # load HIV status data
-hivstatus_vlcopies_raw <- read_hiv_data(path_hiv_614,
-                                        path_hiv_1518,
-                                        path_hiv_19)
+hivstatus_vlcopies_raw <- read_hiv_data(file.path.hiv.614,
+                                        file.path.hiv.1518,
+                                        file.path.hiv_19)
 
 # load HIV verification data set
-verif_raw <- read_flow_data(path_flow_614,
-                            path_flow_1518,
-                            path_flow_19)
+verif_raw <- read_flow_data(file.path.flow.614,
+                            file.path.flow,
+                            file.path.flow_19)
 
 
 ################
@@ -139,17 +111,21 @@ anonimized_id[, anonimized_id := paste0("RkINC-",
 ## 3. Resides in inland communities
 
 # Subset data based inclusion criteria
-inland_comms <- c("1", "2", "4", "5", "6", "7", "8",
-                  "16", "19", "22", "24", "29", "33",
-                  "34", "40", "56", "57", "58", "62",
-                  "74", "77", "89", "94", "106", "107",
-                  "108", "120", "391", "602", "754")
 
-# sensitivity analysis
-# verif_el <- verif_1[resident == 1 &
-#                     ageyrs >= 15 &
-#                     ageyrs <= 49 &
-#                     comm_num %in% inland_comms, ]
+if(restrict_to_30_comms){
+  # sensitivity analysis
+  inland_comms <- c("1", "2", "4", "5", "6", "7", "8",
+                    "16", "19", "22", "24", "29", "33",
+                    "34", "40", "56", "57", "58", "62",
+                    "74", "77", "89", "94", "106", "107",
+                    "108", "120", "391", "602", "754")
+  
+  verif_el <- verif_1[resident == 1 &
+                      ageyrs >= 15 &
+                      ageyrs <= 49 &
+                      comm_num %in% inland_comms, ]
+
+}
 
 # central analysis
 verif_el <- verif_1[resident == 1 &
@@ -235,23 +211,38 @@ for (i in 1:N) {
 ################
 
 # anonimized id
-file <- file.path(indir_deepsequencedata,
-                  "RCCS_data_estimate_incidence_inland_R6_R18",
-                  "220903",
-                  "anonymized_id_for_incidence_estimate_221129.csv")
-write.csv(anonimized_id, file = file, row.names = FALSE)
+file.name <- file.anonymised.id
+if(! file.exists(file.name))
+{
+  cat("\n Saving output file", file.name, "\n")
+  write.csv(anonimized_id, file = file, row.names = FALSE)
+}else{
+  cat("\n Output file", file.name, "already exists\n")
+}
 
 # statistics
-file_name <- file.path(outdir,
-                       "incidence_inland_statistics_cohort_for_paper_221129.rds")
-saveRDS(stats, file_name)
+file_name <- file.path(outdir,"incidence_inland_statistics_cohort_for_paper_221129.rds")
+if(! file.exists(file.name))
+{
+  cat("\n Saving output file", file.name, "\n")
+  saveRDS(stats, file_name)
+}else{
+  cat("\n Output file", file.name, "already exists\n")
+}
 
 # seroconvert cohort central analysis
-file <- file.path(indir, "data", "seroconverter_cohort_R6R19.rds")
-saveRDS(seroconverter_cohort_list, file = file)
+if(!restrict_to_30_comms){
+  file.name <- file.path.seroconverter_cohort
+}else{
+  file.name <- file.path.seroconverter_cohort.30
+}
 
-# seroconvert cohort sensitivity analysis (continuously surveyed commm)
-# file <- file.path(indir,
-#                   "data",
-#                   "seroconverter_cohort_30comm_R6R19.rds")
-# saveRDS(seroconverter_cohort_list, file = file)
+if( !file.exists(file.name))
+{
+  cat('\n Careful: This data should already exist exist in ', file.name  )
+  cat('\n check that your Zenodo path is correctly specified in config.R ' )
+  cat('\nIf you wish to proceed, and save this file anyway run the commented line below')
+  #   saveRDS(seroconverter_cohort_list, file = file)
+}else{
+  cat('\n Output file', file.name,'already exists.\n')
+}
